tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.type.delete.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.canvas.host.type.delete</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.type.delete</h1>
<p class="desc">window.HTMLCanvasElement interface object is [[Configurable]]</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("window.HTMLCanvasElement interface object is [[Configurable]]");
_addTest(function(canvas, ctx) {

  _assertSame(delete window.HTMLCanvasElement, true, "delete window.HTMLCanvasElement", "true");
  _assertSame(window.HTMLCanvasElement, undefined, "window.HTMLCanvasElement", "undefined");

});
</script>
