LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/canvas-host/2d.canvas.host.type.delete-expected.txt

2d.canvas.host.type.delete
window.HTMLCanvasElement interface object is [[Configurable]]
Actual output:

PASS window.HTMLCanvasElement interface object is [[Configurable]]
